watcher-monolith/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
tqdm==4.66.1
python-multipart==0.0.6
pydantic==2.4.2
pytest==7.4.3
httpx==0.25.1
PyPDF2==3.0.1
pdfplumber==0.10.3
sqlalchemy==2.0.27
alembic==1.13.1
aiosqlite==0.19.0
greenlet==3.0.3
aiofiles==23.2.1

# Agentic AI dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
python-socketio==5.11.0
websockets==12.0

